Move getEndOfMonth.js to luxon dir
