+++ app/scripts/actions/就業週報画面に今月ボタンを追加.js
@@ -1,5 +1,5 @@
 import getStartOfMonth from '../modules/luxon/getStartOfMonth'
-import getEndOfMonth from '../modules/getEndOfMonth'
+import getEndOfMonth from '../modules/luxon/getEndOfMonth'
 
 export default function () {
   const searchButton = getSearchButtonElement()
